+++ cmd/xdcd/main.go
@@ -1,0 +1,7 @@
+package words
+
+import "fmt"
+
+func main() {
+	fmt.Println("Hello world")
+}
