+++ go.mod
@@ -4,7 +4,6 @@ go 1.22.0
 
 require (
 	github.com/kljensen/snowball v0.9.0
-	github.com/rookii/paicehusk v0.0.0-20130619054030-d62367ab30c1
 	github.com/tebeka/snowball v0.7.0
 	github.com/toadharvard/stopwords-iso v0.1.2
 )
