+++ pkg/words/utils.go
@@ -1,4 +1,4 @@
-package main
+package words
 
 import (
 	"errors"
