+++ pkg/words/words_test.go
@@ -1,4 +1,4 @@
-package main
+package words
 
 import (
 	"fmt"
